webscraper/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
exceptiongroup==1.1.2
feedparser==6.0.10
Flask==2.3.2
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
selenium==4.10.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
Werkzeug==2.3.6
wsproto==1.2.0
yarl==1.9.2
python-dotenv~=1.0.0
pandas~=2.0.3
spacy~=3.5.3